refactor packages et application class
