+++ src/main/java/filrouge/gedesaft/GedesaftApplication.java
@@ -1,12 +1,12 @@
-package co.simplon.springboot.lesexperts;
+package filrouge.gedesaft;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class LesExpertsApplication {
+public class GedesaftApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(LesExpertsApplication.class, args);
+		SpringApplication.run(GedesaftApplication.class, args);
 	}
 }
