+++ src/main/java/filrouge/gedesaft/model/PersonnelModel.java
@@ -1,4 +1,4 @@
-package co.simplon.springboot.lesexperts.models;
+package filrouge.gedesaft.model;
 
 import java.util.List;
