+++ src/main/java/filrouge/gedesaft/model/VictimeModel.java
@@ -1,4 +1,4 @@
-package co.simplon.springboot.lesexperts.models;
+package filrouge.gedesaft.model;
 
 public class VictimeModel extends ProtagonnisteModelImpl {
